+++ src/kiezellisp-lib/meta.cs
@@ -69,10 +69,12 @@ namespace Kiezel
                     }
                     case "lambda":
                     {
+                        return form;
                         throw new LispException( "lambda not supported by code walker" );
                     }
                     case "letfun":
                     {
+                        return form;
                         throw new LispException( "letfun not supported by code walker" );
                     }
                     case "try":
@@ -85,6 +87,7 @@ namespace Kiezel
                     case "defmulti":
                     case "defmethod":
                     {
+                        return form;
                         throw new LispException( "defun, defmacro, defmulti and defmethod not supported by code walker" );
                     }
                     case "do":